requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.6.0
bcrypt==4.0.1
blinker==1.4
Brlapi==0.8.4
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.5
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu1
Django==4.2.1
djangorestframework==3.14.0
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.103.1
filelock==3.12.0
greenlet==2.0.2
gyp==0.1
h11==0.14.0
httpcore==0.17.2
httpie==3.2.2
httplib2==0.20.4
httpx==0.24.1
idna==3.4
importlib-metadata==4.6.4
jeepney==0.8.0
keyring==23.8.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.22.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
passlib==1.7.4
pexpect==4.8.0
Pillow==9.2.0
platformdirs==3.5.1
psycopg2==2.9.6
ptyprocess==0.7.0
pyasn1==0.5.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.4.0
pydantic_core==2.10.0
Pygments==2.16.1
PyGObject==3.42.2
PyJWT==2.4.0
pyparsing==3.0.7
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
python-apt==2.4.0+22.10
python-dateutil==2.8.1
python-debian==0.1.46+ubuntu1
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-telegram==0.18.0
python-telegram-bot==20.3
pytz==2022.2.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.11
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.6.0
rsa==4.9
SecretStorage==3.3.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
starlette==0.27.0
systemd-python==235
telebot==0.0.5
telegram-text==0.1.1
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.2
uvicorn==0.23.2
vboxapi==1.0
virtualenv==20.23.0
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
